Improve placement of planets/sputniks
Previously, the values passed as ‘clearance’ to
Galaxy#random_location weren’t ideal. Planets
and their orbits could be fully contained within
another planet’s orbit which would mean it’s
impossible to visit orbits according to planets’
ordinals.

Additionally, the ‘clearance’ for sputniks was
wrong and should be based on the bounding circle
of the hull rather than the hull area scalar.